orchestrator/requirements/base.txt
# Core dependencies for Binks Orchestrator
crewai>=0.1.0
crewai-tools>=0.1.0
langchain-community>=0.0.1
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0
python-dotenv>=1.0.0
kubernetes>=28.0.0
pyyaml>=6.0.0
requests>=2.31.0

# Optional but recommended
rich>=13.0.0  # For better console output
loguru>=0.7.0  # For logging
